move the interaction patterns into separate modules
