+++ src/pattern/mod.rs
@@ -1,2 +1,6 @@
 //!
+pub mod bidi_streaming;
+pub mod client_streaming;
+pub mod rpc;
+pub mod server_streaming;
 pub mod try_server_streaming;
